Remove unused getConfigSpec methods from ApiDemos sample code
At one time during the development of GLSurfaceView, the
GLSurfaceView.Renderer interface had a getConfigSpec method that was used to
select the frame buffer format. This method was removed from the
Renderer interface before GLSurfaceView became a public API, but
the sample code still defined these now-unused methods.

To reduce the chance of confusing a developer who uses the sample code,
this change removes the dead, no-longer-called getConfigSpec
implementations from the three samples.

See GLSurfaceView.setEGLConfigChooser for the replacement
technique.