Modfied demo to show setting animations to null
Originally, setting LayoutTransition animators to null just restored the default animation behavior, It seems more sensible to have null actually disable the animation. I made that change to the framework and modified this demo to show how it's done. Along the way, I refactored the code a bit and fixed some minor bugs.

Change-Id: Icfb53ada234b635294c5faa68a15932f25859fb8