Improve TransportController demo.
This is an actually realistic implementation of a
video player.  It's still delibrately simple, but it's
not outright broken.  We now have our own media controller
since that is the only way to get something that actually
works, we use system UI flags to hide nav/status bar when
playing, we use the new TransportController APIs to go
into a formal playback state where we take audio focus.

Change-Id: I19ac171483637b8ee94c9c07aea39c11748cbc46