Improve photos content observer demo.
This is now much more like one would write to look for
photos being added/changed.  It looks at the URIs that have
changed and filters them to only files that it cares about
(in this case those under DCIM, that is photos added by the
camera), and also explicitly understands the difference
between individual URIs changing and a more general change
requiring a full re-query of the provider.

Change-Id: I509570aa66839f8b84b5a065c4b4e49fbd0f1c29